+++ app/adapters/memory.py
@@ -1,14 +1,17 @@
-"""A memory driven adapter for data storage."""
+"""A memory database adapter."""
 
-from ..domain import Book, BookRepository, BookView
-from ..controller import UnitOfWorkManager, UnitOfWork
+from ..settings import identify
+from ..domain.model import Book
+from ..domain.ports import BookRepository, BookView, UnitOfWork, \
+						   UnitOfWorkManager
 
 
 BOOKS_STORAGE = {}
 
 
 class MemoryBookRepository(BookRepository):
-	"""An implementation of a BookRepository with memory storage.
+	"""An implementation of a BookRepository utilizing memory storage as a
+	database for the application.
 
 	Methods: save
 	"""
@@ -17,7 +20,7 @@ class MemoryBookRepository(BookRepository):
 		pass
 
 	def save(self, book: Book):
-		"""View @domain.BookRepository."""
+		"""View @app.domain.ports.BookRepository."""
 		BOOKS_STORAGE[book.isbn] = {'name': book.name, 'author': book.author,
 									'content': book.content}
 
@@ -32,12 +35,12 @@ class MemoryBookView(BookView):
 		pass
 
 	def get_all(self) -> list:
-		"""View @domain.BookView."""
+		"""View @app.domain.ports.BookView."""
 		return [Book(key, value['name'], value['author'], value['content']) \
 				for key, value in BOOKS_STORAGE.items()]
 
 	def get_by_isbn(self, isbn: str) -> Book:
-		"""View @domain.BookView."""
+		"""View @app.domain.ports.BookView."""
 		book = BOOKS_STORAGE.get(isbn)
 
 		if book is not None:
@@ -46,7 +49,7 @@ class MemoryBookView(BookView):
 			return None
 
 	def get_by_name(self, name: str) -> list:
-		"""View @domain.BookView."""
+		"""View @app.domain.ports.BookView."""
 		books = []
 		for key, value in BOOKS_STORAGE.items():
 			if value['name'] == name:
@@ -57,7 +60,7 @@ class MemoryBookView(BookView):
 		return books
 
 	def get_by_author(self, author: str) -> list:
-		"""View @domain.BookView."""
+		"""View @app.domain.ports.BookView."""
 		books = []
 		for key, value in BOOKS_STORAGE.items():
 			if value['author'] == author:
@@ -69,7 +72,7 @@ class MemoryBookView(BookView):
 
 
 class MemoryUnitOfWork(UnitOfWork):
-	"""An implementation of a UnitOfWork for memory storage.
+	"""An implementation of a UnitOfWork for a memory storage database.
 
 	Methods: __enter__, __exit__, commit, rollback, books
 	"""
@@ -78,24 +81,24 @@ class MemoryUnitOfWork(UnitOfWork):
 		pass
 
 	def __enter__(self):
-		"""View @controller.UnitOfWork."""
+		"""View @app.domain.ports.UnitOfWork."""
 		return self
 
 	def __exit__(self, type, value, traceback):
-		"""View @controller.UnitOfWork."""
+		"""View @app.domain.ports.UnitOfWork."""
 		pass
 
 	def commit(self):
-		"""View @controller.UnitOfWork."""
+		"""View @app.domain.ports.UnitOfWork."""
 		pass
 
 	def rollback(self):
-		"""View @controller.UnitOfWork."""
+		"""View @app.domain.ports.UnitOfWork."""
 		pass
 
 	@property
 	def books(self) -> BookRepository:
-		"""View @controller.UnitOfWork."""
+		"""View @app.domain.ports.UnitOfWork."""
 		return MemoryBookRepository()
 
 
@@ -108,6 +111,39 @@ class MemoryUnitOfWorkManager(UnitOfWorkManager):
 		"""MemoryUnitOfWorkManager's constructor."""
 		pass
 
-	def start(self) -> UnitOfWork:
-		"""View @controller.UnitOfWorkManager."""
+	def start(self) -> MemoryUnitOfWork:
+		"""View @app.domain.ports.UnitOfWorkManager."""
 		return MemoryUnitOfWork()
+
+
+@identify('memory', 'database')
+class MemoryAdapter(object):
+	"""The adapter class to be used by the application. It gives access to
+	each of the memory database classes that are to be used by the app.
+
+	Methods: get_uowm, get_view
+	"""
+	def __init__(self, cfg: dict):
+		"""MemoryAdapter's constructor.
+
+		cfg: dict -- The memory adapter's configuration
+		"""
+		pass
+
+	def get_uowm(self):
+		"""Returns an instance of a MemoryUnitOfWorkManager.
+
+		Returns
+		-------
+		uowm: MemoryUnitOfWorkManager
+		"""
+		return MemoryUnitOfWorkManager()
+
+	def get_view(self):
+		"""Returns an instance of a MemoryBookView.
+
+		Returns
+		-------
+		view: MemoryBookView
+		"""
+		return MemoryBookView()